server/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
dnspython
